mcp: support stateless connections for streamable
When the client lost the connection to the mcp server or the mcp server runs on distributed nodes (e.g. behind a loadbalancer), it can be desired to run the server in stateless mode. When doing so, the Mcp-Session-Id should not be verified to allow the execution of methods without prior initialization.

Fixes #148